Complete Assignment 42

+++ app/controllers/comments_controller.rb
@@ -2,30 +2,61 @@ class CommentsController < ApplicationController
   before_action :require_sign_in
   before_action :authorize_user, only: [:destroy]
 
-  def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
+  def show
+    @comment = Comment.find(params[:id])
+  end
 
-    if comment.save
-      flash[:notice] = 'Comment saved successfully.'
-      redirect_to [@post.topic, @post]
+  def create
+    if params[:post_id].present?
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = 'Comment saved successfully.'
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = 'Comment failed to save.'
+        redirect_to [@post.topic, @post]
+      end
     else
-      flash[:alert] = 'Comment failed to save.'
-      redirect_to [@post.topic, @post]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = 'Comment saved successfully.'
+        redirect_to @topic
+      else
+        flash[:alert] = 'Comment failed to save.'
+        redirect_to @topic
+      end
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
+    if params[:post_id].present?
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.find(params[:id])
 
-    if comment.destroy
-      flash[:notice] = 'Comment was deleted successfully.'
-      redirect_to [@post.topic, @post]
+      if comment.destroy
+        flash[:notice] = 'Comment was deleted successfully.'
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic, @post]
+      end
     else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic, @post]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.find(params[:id])
+
+      if comment.destroy
+        flash[:notice] = 'Comment was deleted successfully.'
+        redirect_to @topic
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to @topic
+      end
     end
   end
 
@@ -36,10 +67,16 @@ class CommentsController < ApplicationController
   end
 
   def authorize_user
-    comment = Comment.find(params[:id])
-    unless current_user == comment.user || current_user.admin?
+    @comment = Comment.find(params[:id])
+    @post = Post.find(params[:post_id])
+    # @topic = Topic.find(params[:topic_id])
+    unless current_user == @comment.user || current_user.admin?
       flash[:alert] = 'You do not have permission to delete a comment.'
-      redirect_to [comment.post.topic, comment.post]
+      if params[:post_id].present?
+        redirect_to root_path
+      else
+        redirect_to root_path
+      end
     end
   end
 end
